grpc_routes_analysi/requirements.txt
flake8==6.0.0
grpcio==1.51.1
grpcio-tools==1.51.1
mccabe==0.7.0
protobuf==4.21.12
pycodestyle==2.10.0
pyflakes==3.0.1
joblib==1.2.0
numpy==1.23.5
scikit-learn==1.2.0
scipy==1.9.3
threadpoolctl==3.1.0
xgboost==1.7.2
